Added roles.Update functionality and moved common code
