+++ internal/roles/roles.go
@@ -1,11 +1,8 @@
 package roles
 
 import (
-	"encoding/json"
-
 	p "github.com/TanmoySG/wunderDB/internal/privileges"
 	"github.com/TanmoySG/wunderDB/model"
-	"github.com/TanmoySG/wunderDB/pkg/utils/maps"
 	er "github.com/TanmoySG/wunderDB/pkg/wdb/errors"
 )
 
@@ -32,7 +29,7 @@ func (r Roles) CheckIfExists(roleID model.Identifier) (bool, *model.Role) {
 	}
 }
 
-func (r Roles) CreateRole(roleID model.Identifier, allowed []string, denied []string, hidden bool) *er.WdbError {
+func (r Roles) Create(roleID model.Identifier, allowed []string, denied []string, hidden bool) *er.WdbError {
 
 	grants, err := getPrivileges(allowed, denied)
 	if err != nil {
@@ -53,7 +50,7 @@ func (r Roles) CreateRole(roleID model.Identifier, allowed []string, denied []st
 
 }
 
-func (r Roles) ListRoles(forceListAllRoles bool) Roles {
+func (r Roles) List(forceListAllRoles bool) Roles {
 	var filteredRoles = make(Roles)
 
 	if forceListAllRoles {
@@ -128,100 +125,18 @@ func (r Roles) Check(permissions []model.Permissions, privilege string, on *mode
 	return denied
 }
 
-func checkPermission(privilege string, rolePrivileges model.Privileges) bool {
-	privilegeAllowed, privilegeExists := rolePrivileges[privilege]
-	if privilegeExists {
-		if privilegeAllowed == p.Allowed {
-			return p.Allowed
-		}
-	}
-	return p.Denied
-}
-
-func getPrivileges(allowedActions, deniedActions []string) (*model.Grants, error) {
-
-	var grants model.Grants
-
-	allowedGrantsMap := sortPrivileges(allowedActions, p.Allowed)
-	deniedGrantsMap := sortPrivileges(deniedActions, p.Denied)
-
-	userPrivileges, err := mergeGrantMaps(maps.Marshal(allowedGrantsMap.UserPrivileges), maps.Marshal(deniedGrantsMap.UserPrivileges))
+func (r Roles) Update(roleID model.Identifier, allowed []string, denied []string, hidden bool) *er.WdbError {
+	grants, err := getPrivileges(allowed, denied)
 	if err != nil {
-		return nil, err
-	}
-
-	globalPrivileges, err := mergeGrantMaps(maps.Marshal(allowedGrantsMap.GlobalPrivileges), maps.Marshal(deniedGrantsMap.GlobalPrivileges))
-	if err != nil {
-		return nil, err
-	}
-
-	databasePrivileges, err := mergeGrantMaps(maps.Marshal(allowedGrantsMap.DatabasePrivileges), maps.Marshal(deniedGrantsMap.DatabasePrivileges))
-	if err != nil {
-		return nil, err
-	}
-
-	collectionPrivileges, err := mergeGrantMaps(maps.Marshal(allowedGrantsMap.CollectionPrivileges), maps.Marshal(deniedGrantsMap.CollectionPrivileges))
-	if err != nil {
-		return nil, err
-	}
-
-	grants.GlobalPrivileges = globalPrivileges
-	grants.DatabasePrivileges = databasePrivileges
-	grants.CollectionPrivileges = collectionPrivileges
-	grants.UserPrivileges = userPrivileges
-
-	return &grants, nil
-
-}
-
-func sortPrivileges(actions []string, assignedPermission bool) model.Grants {
-	var privilegeGrants model.Grants
-
-	userPrivileges := model.Privileges{}
-	globalPrivileges := model.Privileges{}
-	databasePrivileges := model.Privileges{}
-	collectionPrivileges := model.Privileges{}
-
-	for _, action := range actions {
-		if p.IsAvailable(action) {
-			actionCategory := p.Category(action)
-			switch actionCategory {
-			case p.DatabasePrivileges:
-				databasePrivileges[action] = assignedPermission
-			case p.CollectionPrivileges:
-				collectionPrivileges[action] = assignedPermission
-			case p.GlobalPrivileges:
-				globalPrivileges[action] = assignedPermission
-			case p.UserPrivileges:
-				userPrivileges[action] = assignedPermission
-			}
+		return &er.WdbError{
+			ErrCode:        er.EncodeDecodeError.ErrCode,
+			ErrMessage:     err.Error(),
+			HttpStatusCode: er.EncodeDecodeError.HttpStatusCode,
 		}
 	}
 
-	privilegeGrants.UserPrivileges = &userPrivileges
-	privilegeGrants.GlobalPrivileges = &globalPrivileges
-	privilegeGrants.DatabasePrivileges = &databasePrivileges
-	privilegeGrants.CollectionPrivileges = &collectionPrivileges
+	r[roleID].Grants = *grants
+	r[roleID].Hidden = hidden
 
-	return privilegeGrants
-}
-
-func mergeGrantMaps(allowedPrivilegesMap, deniedPrivilegesMap map[string]interface{}) (*model.Privileges, error) {
-	var privileges model.Privileges
-
-	mergedGrantsMap, err := maps.Merge(allowedPrivilegesMap, deniedPrivilegesMap)
-	if err != nil {
-		return nil, err
-	}
-
-	mergedGrantBytes, err := json.Marshal(mergedGrantsMap)
-	if err != nil {
-		return nil, err
-	}
-
-	err = json.Unmarshal(mergedGrantBytes, &privileges)
-	if err != nil {
-		return nil, err
-	}
-	return &privileges, nil
+	return nil
 }
